float

+++ 0x00-python_variable_annotations/5-sum_list.py
@@ -3,10 +3,10 @@
 Module for Type Annotation
 """
 
-type Input_list = List[int]
+type Input_list = List[float]
 
 
-def sum_list(input_list: Input_list) -> int:
+def sum_list(input_list: Input_list) -> float:
     """
     Function that takes a list of integers as argument and returns their sum
     """
